Правки замечений
