+++ Movie/Movie/Builder/Builder.swift
@@ -7,16 +7,16 @@ import UIKit
 final class Builder: BuilderProtocol {
     // MARK: - Private properties
 
-    private let networkManager = NetworkManager()
-    private let imageAPIService = ImageAPIServiceImp()
+    private let networkService = NetworkService()
+    private let imageAPIService = ImageAPIService()
     private let cacheService = CacheService()
 
     // MARK: - Public methods
 
     func buildCinemaListScreen() -> UIViewController {
         let proxy = ProxyImp(imageAPIService: imageAPIService, cacheService: cacheService)
-        let imageService = ImageServiceImp(proxy: proxy)
-        let viewModel = CinemaListScreenViewModelImp(networkManager: networkManager, imageService: imageService)
+        let imageService = ImageService(proxy: proxy)
+        let viewModel = CinemaListScreenViewModel(networkManager: networkService, imageService: imageService)
         let internalView = CinemaListInternalView(frame: CGRect(x: 0, y: 0, width: 100, height: 100))
         let viewController = CinemaListViewController()
         viewController.internalView = internalView
@@ -26,8 +26,8 @@ final class Builder: BuilderProtocol {
 
     func buildCinemaDescriptionScreen(cinema: Cinema, imageData: Data) -> UIViewController {
         let proxy = ProxyImp(imageAPIService: imageAPIService, cacheService: cacheService)
-        let imageService = ImageServiceImp(proxy: proxy)
-        let viewModel = CinemaDescriptionViewModelImp(networkService: networkManager, imageService: imageService)
+        let imageService = ImageService(proxy: proxy)
+        let viewModel = CinemaDescriptionViewModelImp(networkService: networkService, imageService: imageService)
         let viewController = CinemaDescriptionViewController(cinema: cinema, imageData: imageData)
         viewController.viewModel = viewModel
         return viewController
